chugging along, added errors structs for when a bad request is sent, need to do better error handling, still implementing retrieving manga volume info etc
